Integração cadastro de produtos

+++ src/components/AppContainer.js
@@ -1,7 +1,6 @@
 import React, {Component} from 'react'
 import 'typeface-roboto'
 import axios from "axios"
-import CadastroProduto from './CadastroProduto/CadastroProduto'
 import { LandingPage } from './LandingPage/LandingPage'
 
 
@@ -23,17 +22,21 @@ export class AppContainer extends Component {
   }
 
   addProduct = (body) => {
-    axios.post(url, body).then(response => console.log(response.menssage)).catch(err => console.log(err.menssage))
+    axios.post(url, body).then(response => console.log(response.data)).catch(err => console.log(err.menssage))
+    this.getAllProducts()
   }
   
   delProduct = id => {
-    axios.delete(url + "/" + id)
+    if (window.confirm("Tem certeza que quer excluir esse produto?")) {
+      axios.delete(url + "/" + id)
+      this.getAllProducts()
+    }
   }
   
     render() {
       return ( 
       <div>
-        <LandingPage produtos={this.state.products}/>
+        <LandingPage produtos={this.state.products} reload={this.getAllProducts} addProduto={this.addProduct} delProduct={this.delProduct}/>
         </div>
       )
     }
